Added Equals and GetHashCode to classes we use as keys in Dictionaries and HashSets
Added TileHeightChanged event to Unit.

Added propagation of MovementType through every dist calc method, visitors etc.

Added NodeEqualityComparer to IPathFindAlg to enable easier usage in INode as a key in Dictionary

Changed getting closestNode search to prefer BuildingNodes, so if there is TileNode and BuldingNode at the same spot, it returns the BuildingNode

Changed checking of passability to not use TempNodes, but backup or go forward to TIleNode or BuildingNode

Fixed MapFormationController to stop trying to order units of the same type to the same tile after the first one fails

Fixed BFS roof formation controller to properly order the units to different nodes

Implemented unit movement.