Changed MeeleAttacker to enable implementation of attack move. - also to move immedietely on attack command instead of wayting for position check. - changed to lose target when movingattacker cannot get to the target
Added GameUI.RegisterForHover and GameUI.UnregisterForHover to enable independent registration for hover detection, so that even plugin defined UI can be tracked and disable Tools when hovering.
- added calls to registration and unregistration to every tool and plugin adding UI elements

Changed Building building and CanBuild checking to just take top left corner instead of the whole rectangle
- because in knows the size
- added API to get the rectangle from topLeft corner

Added visibility toggle to HealthBar

Changed ControllerFactory from abstract class to interface, to be in line with other parts of Input

Fixed Building loading, which used to create the Building component AFTER it tried to add it onto a node

Added tile height change checking with the building on the tile. Now it is routed to the plugin and we ask if the heightcan be changed. Then there is a notification that the height has changed.

Added comments to lots of things, need much much more.

Removed LegNode and CenterNode distinction from Unit, it was remnant of ancient history which was not used anymore.

Changed loading again, for the final time. Added explicit posting of the method calls onto the Scheduler made from the synchronization context, so all loading is now done on the main thread.

Added Player removal from game. If we remove human player, it triggers the end of the game.

Fixed TryGetXXX from node, which now traverses the parental chain to find the root node of the entity or levelNode or scene node, whatever it finds.

Added unit notification on building destruction on the same tile the unit is standing on

Renamed LevelRep states to match the documentation

Added the ability to change BuildingNode position in the Pathfinding algorithm

Fixed LevelCreationScreen and other screens that were throwing out the provided data on Hide(), which then on Back() call ment that they did not have any data to show

Added additional parameter to GetTilesInSpiral named cutoff which limits the size of the spiral returned and so makes it finite enumerable instead of infinite. It is optinal parameter, default call is still infinite

Fixed MapGraphics Dispose call which could not handle disposing on loading failure, because it did not have everything loaded.

Added PackageDefinition template file.

Added Package UI layout, which adds top bar to show resources on

Added check to CanBuild of buildings to check the difference in tile height is low enough to build

Fixed Formation controllers to abort ordering units when one unit of the type fails

Fixed memory leak, where we again did not remove all handlers from the UI

Fixed healthBarCreation and loading, fixed UI loading on Gate and Keep

Implemented AggressivePlayer AI, need to test it

Added additional implementations of DistCalc to chicken and wolf to easily compute the path thorough enemy buildings that they can break