Fixed Dog behavior, fixed Gate and Keep UI style not applying (not a nice fix, with time could be fixed better),  added style to TopBar of plugin UI, fixed treeBuilder not respecting UI.Hovering, finished and removed some TODOs
Fixed dog and treeCutter behavior where they had really expensive or even infinite loops, inifite recursion and other nice things. Moved most of the computation to OnUpdate, because events created the infinite recursions.

Fixed UI style application by changing the way the UI is loaded. Now it loads it together with the style file, instead of setting the style file to pluginUI root. This is a hack because i cant make the style at root work. It also forced me to wrap the window elements in another UIElement, otherwise it would be created on load.

Removed implementation of Android file manager, to not scare people off.

Removed some TODOs that were left there after the fixes.

Changed some TODOs to FUTURE tags

Removed some TODOs as i dont plan to fix them and they work as they are

Changed SaveAs screen to check if the Level it is saving is set.

Removed all use of CursorTooltips that were never finished and never used. Maybe in the future.

Changed MovingRangeTarget and MovingMeeleAttacker to require to be placed on a Unit.