Changed some TODOs to FUTURE and ALT.
Fixed error that was thrown when Shooters were enumerated and deleted on Target death and they unregistered themselfs, changing the IEnumerable and causing an exception. Now we create a copy before enumeration.

Renamed all thinks containing MandK to MouseKeyboard

Fixed missing player plugin disposal

Removed IDisposable from resource, there is nothing to dispose

Added the disposal of Package on error during loading, also added SendFailed message

Added error checking to LevelRep XML loading.

Changed AStar visualization so that it can be set from the options screen during the run of the game.

 Added HasEdgeTo to INode so we could implement the pathfinding graph creation without duplicating the edges

Added Pathfinding visualization to Options and options screen

Fixed Expanding selector checking and unchecking without changing the state of tool or player. Disabled the selected checkbox so that it cannot be unchecked.

Fixed SaveAsScreen to really save a copy of the current level, without switching the levelRep so that it is more like a snapshot.

Added missing mono-urho.dll to Desktop project, dont know where they are comming from in the output, but now they can be copied into the installer.

Added CoreData.pak to copy it into the installer.

Fixed pathfinding node creation, moved it to LoadState and CreateNew so that all the things are already loaded when creating the graph.

Fixed duplication of the edges in the pathfinding graph.

Changed damage to buildings. Now they can only be damaged by projectiles.

Fixed lazy player to end the level when his keep is destroyed.